refactor: replaced _FinalizationQueue
The _FinalizationQueue class was originally intended to finalize more
than just Device instances, and as such, it was important to preserve
the insertion order to ensure proper resource cleanup. However, it was
never implemented in this way, and so it is suitable to use a
WeakKeyDictionary instead.

Also tweaked some messages and docstrings.